Update services_eo.properties
It... seems a bit weird, because I can't quite tell which instances of gxi are "dummy it"s. My intuition is saying it's safer to cut those out.
I'm not sure why "amuze" is there. Are you maybe looking for "Nu"? (It's kind of a cross between hmm and pardonon in this context imo.)
While enigi is fine, entajpi is another option (style).