Update tree-services_eo.properties
It's just style, but you could change it to "totalo de" instead of "totala nombro de" for brevity.
"sola" seems a teensy bit weird here but I'm not sure why or what might replace it.
Changed "sciarbo" to "scioarbo" to make it more in line with the stats. You could possibly change stats to "Scioarbo". You might have to check what the "Printi la Scioarbon" was in the other section?
Once again, "entajpu" is an alternative to "enigu" if wanted.
I've just thought that the other section might have "statistikoj" instead of "statistiko", that could be changed to match this.