Update game_eo.properties
Ach, I forgot that line! Just changed a verb. Also: I haven't checked a technical dictionary but I'm not sure if "enen" is the right word for the enter key itself. Also, the top says "You're thinking of an animal and I'll guess it". If it's meant to be "Think of...", "pensu" is better.